+++ src/automation/task-executor.ts
@@ -102,8 +102,11 @@ const OPERATION = {
  */
 export class TaskExecutor implements Composable {
     taskRefs = {};
+    maxModelArraySize = 10;
 
     initialize(): Composable {
+        const config = AppConfig.getInstance();
+        this.maxModelArraySize = util.str2int(config.getProperty("max.model.array.size", "1000"));
         return this;
     }
 
@@ -228,8 +231,8 @@ export class TaskExecutor implements Composable {
         for (const entry of mapping) {
             const sep = entry.indexOf(MAP_TO);
             if (sep > 0) {
-                let lhs = entry.substring(0, sep).trim();
-                const rhs = entry.substring(sep+2).trim();
+                let lhs = this.substituteDynamicIndex(entry.substring(0, sep).trim(), source, false);
+                const rhs = this.substituteDynamicIndex(entry.substring(sep+2).trim(), source, true);
                 const isInput = lhs.startsWith(INPUT_NAMESPACE) || lhs.toLowerCase() == INPUT;
                 if (lhs.startsWith(INPUT_HEADER_NAMESPACE)) {
                     lhs = lhs.toLowerCase();
@@ -403,10 +406,10 @@ export class TaskExecutor implements Composable {
         for (const entry of mapping) {
             const sep = entry.indexOf(MAP_TO);
             if (sep > 0) {
-                const lhs = entry.substring(0, sep).trim();
+                const lhs = this.substituteDynamicIndex(entry.substring(0, sep).trim(), consolidated, false);
                 const isInput = lhs.startsWith(INPUT_NAMESPACE) || lhs.toLowerCase() == INPUT;
                 let value = null;
-                const rhs = entry.substring(sep+2).trim();
+                const rhs = this.substituteDynamicIndex(entry.substring(sep+2).trim(), consolidated, true);
                 if (isInput || lhs.startsWith(MODEL_NAMESPACE)
                         || lhs == HEADER || lhs.startsWith(HEADER_NAMESPACE)
                         || lhs == STATUS
@@ -906,6 +909,44 @@ export class TaskExecutor implements Composable {
             return this.getValueByType(type, value, "LHS '"+lhs+"'", source);            
         }
         return value;
+    }
+
+    private substituteDynamicIndex(text: string, source: MultiLevelMap, isRhs: boolean): string {
+        if (text.includes('[model.')) {
+            let sb = '';
+            let start = 0;
+            while (start < text.length) {
+                const open = text.indexOf('[', start);
+                const close = text.indexOf(']', start);
+                if (open != -1 && close > open) {
+                    sb += text.substring(start, open+1);
+                    const idx = text.substring(open+1, close).trim();
+                    if (idx.startsWith(MODEL_NAMESPACE)) {
+                        const ptr = util.str2int(String(source.getElement(idx)));
+                        if (isRhs) {
+                            if (ptr > this.maxModelArraySize) {
+                                throw new Error("Cannot set RHS to index > " + ptr
+                                        + " that exceeds max "+this.maxModelArraySize+" - "+text);
+                            }
+                            if (ptr < 0) {
+                                throw new Error("Cannot set RHS to negative index - " + text);
+                            }
+                        }
+                        sb += String(ptr);
+                    } else {
+                        sb += idx;
+                    }
+                    sb += ']';
+                    start = close + 1;
+                } else {
+                    sb += text.substring(start);
+                    break;
+                }
+            }
+            return sb;
+        } else {
+            return text;
+        }
     }
 
     private getValueByType(type: string, value, path: string, data: MultiLevelMap) {